Fix up bad copy/paste from search to autocomplete
