+++ api/query/autocomplete.go
@@ -37,6 +37,11 @@ type byQueryIndex struct {
 func (r byQueryIndex) Len() int      { return len(r.rs) }
 func (r byQueryIndex) Swap(i, j int) { r.rs[i], r.rs[j] = r.rs[j], r.rs[i] }
 func (r byQueryIndex) Less(i, j int) bool {
+	a := strings.Index(r.rs[i].String, r.q)
+	b := strings.Index(r.rs[j].String, r.q)
+	if a == b {
+		return r.rs[i].String < r.rs[j].String
+	}
 	return strings.Index(r.rs[i].String, r.q) < strings.Index(r.rs[j].String, r.q)
 }
