+++ api/query/autocomplete_test.go
@@ -1,0 +1,82 @@
+// +build small
+
+// Copyright 2018 The WPT Dashboard Project. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+package query
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/web-platform-tests/wpt.fyi/shared"
+)
+
+func TestPrepareAutocompleteResponse_none(t *testing.T) {
+	runIDs := []int64{1, 2}
+	testRuns := []shared.TestRun{
+		shared.TestRun{
+			ID:         runIDs[0],
+			ResultsURL: "https://example.com/1-summary.json.gz",
+		},
+		shared.TestRun{
+			ID:         runIDs[1],
+			ResultsURL: "https://example.com/2-summary.json.gz",
+		},
+	}
+	filters := shared.QueryFilter{
+		RunIDs: runIDs,
+		Q:      "/appears_nowhere/",
+	}
+	summaries := []summary{
+		map[string][]int{
+			"/a/b/c": []int{1, 2},
+			"/b/c":   []int{9, 9},
+		},
+		map[string][]int{
+			"/z/b/c": []int{0, 8},
+			"/x/y/z": []int{3, 4},
+			"/b/c":   []int{5, 9},
+		},
+	}
+
+	resp := prepareAutocompleteResponse(&filters, testRuns, summaries)
+	assert.Equal(t, []AutocompleteResult{}, resp.Results)
+}
+
+func TestPrepareAutocompleteResponse_several(t *testing.T) {
+	runIDs := []int64{1, 2}
+	testRuns := []shared.TestRun{
+		shared.TestRun{
+			ID:         runIDs[0],
+			ResultsURL: "https://example.com/1-summary.json.gz",
+		},
+		shared.TestRun{
+			ID:         runIDs[1],
+			ResultsURL: "https://example.com/2-summary.json.gz",
+		},
+	}
+	filters := shared.QueryFilter{
+		RunIDs: runIDs,
+		Q:      "/b/",
+	}
+	summaries := []summary{
+		map[string][]int{
+			"/a/b/c": []int{1, 2},
+			"/b/c":   []int{9, 9},
+		},
+		map[string][]int{
+			"/z/b/c": []int{0, 8},
+			"/x/y/z": []int{3, 4},
+			"/b/c":   []int{5, 9},
+		},
+	}
+
+	resp := prepareAutocompleteResponse(&filters, testRuns, summaries)
+	assert.Equal(t, []AutocompleteResult{
+		AutocompleteResult{"/b/c"},
+		AutocompleteResult{"/a/b/c"},
+		AutocompleteResult{"/z/b/c"},
+	}, resp.Results)
+}
